Add world generation and states system
The world gen is kinda bad at the moment,
but I'm committing now to save it before I make it simpler.

Improved the person pathfinding.

Made the ship a scene instead of a container, and added the
ship ui scene.

Added the map and test map scenes.

Added the state manager.